FIX: Make sidebars appear on each page.
With newer Sphinx versions, it becomes harder to control the sidebar
behaviour without touching the theme itself. I had to remove the
`hidden` directive for the toctree and tweak the conf file in order to
make the TOC appear on each page. Cost of this: TOC also appears at the
bottom of the index page. Maybe we find a better way someday.